feat: add gitlab support, still working on config options
