+++ lua/nvim-git-line.lua
@@ -1,8 +1,8 @@
---[[ Copyright © 2022 Ryan Ciehanski ]]
+--[[ Copyright © 2022 Ryan Ciehanski <ryan@ciehanski.com> ]]
 --[[ MIT license, see LICENSE for more details. ]]
 
 -- @module "nvim-git-line.github"
-local github = require("nvim-git-line.github")
+local repo = require("nvim-git-line.repo")
 
 local M = {}
 
@@ -21,14 +21,14 @@ function M.setup(user_config)
   -- Set keymappings
   -- Open buffer in remote repo
   api.nvim_set_keymap("n", M.config.action_key, "", {
-    callback = github.go,
+    callback = repo.go,
     desc = "Opens the current file in the GitHub repository with your default browser.",
     silent = not M.config.debug,
     noremap = false,
   })
   -- Open buffer + current line number in remote repo
   api.nvim_set_keymap("n", M.config.action_key_line, "", {
-    callback = github.go_line,
+    callback = repo.go_line,
     desc = "Opens the current file at the line under your cursor in the GitHub repository with your default browser.",
     silent = not M.config.debug,
     noremap = false,
