+++ lua/nvim-git-line/utils.lua
@@ -1,4 +1,4 @@
---[[ Copyright © 2022 Ryan Ciehanski ]]
+--[[ Copyright © 2022 Ryan Ciehanski <ryan@ciehanski.com> ]]
 --[[ MIT license, see LICENSE for more details. ]]
 
 local M = {}
@@ -6,7 +6,7 @@ local M = {}
 local api = vim.api
 local fn = vim.fn
 
--- @param url String
+-- @param url string
 function M.open_url(url)
   -- Open URL in default browser
   local vim_os = vim.loop.os_uname().sysname
@@ -19,7 +19,7 @@ function M.open_url(url)
   end
 end
 
--- @return String
+-- @return string
 function M.get_branch_name()
   local branch = ""
   if M.is_git_dir() then
@@ -34,21 +34,19 @@ function M.get_branch_name()
   return branch
 end
 
--- @return String
+-- @return string
 function M.get_remote_repo_name()
   local repo_name = ""
   if M.is_git_dir() then
     local remote_origin = fn.system("git remote get-url origin | tr -d '\n'")
     local repo_git = remote_origin:sub(remote_origin:find("/[^/]*$") + 1)
-    for v in string.gmatch(repo_git, "[^.]+") do
-      repo_name = v
-      break
-    end
+    local dot_index, _ = repo_git:find(".", 1, true)
+    repo_name = repo_git:sub(0, dot_index - 1)
   end
   return repo_name
 end
 
--- @return String
+-- @return string
 function M.get_remote_username()
   local username = ""
   if M.is_git_dir() then
@@ -59,21 +57,23 @@ function M.get_remote_username()
   return username
 end
 
--- @return String
-function M.get_buffer_name()
+-- @param repo_name string
+-- @return string
+function M.get_buffer_name(repo_name)
   -- Get buffer name from bufnr
   local bufpath = fn.expand("%:p")
-  local bufname = bufpath:sub(bufpath:find("[^/]+/[^/]+/[^/]*$"))
+  local _, eindex = bufpath:find(repo_name)
+  local bufname = bufpath:sub(eindex + 2)
   -- Ship it
   return bufname
 end
 
--- @return String
+-- @return string
 function M.get_line_number()
   return vim.inspect(api.nvim_win_get_cursor(0)[1])
 end
 
--- @return Boolean
+-- @return boolean
 function M.is_git_dir()
   if fn.system("git rev-parse --is-inside-work-tree | tr -d '\n'") == "true" then
     return true
